storage: move state helpers to own file
